refactor(PaymentInfo): 클래스 위치 변경(domain->dto)
