+++ src/main/java/roomescape/dto/PaymentInfo.java
@@ -1,0 +1,4 @@
+package roomescape.dto;
+
+public record PaymentInfo(Long amount, String orderId, String paymentKey) {
+}
